+++ src/main/java/A1BnB/backend/domain/post/dto/response/PostResponse.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 // 게시물 응답
 public record PostResponse(Long postId, String authorName, List<String> photoUrls, String location,
-                           Double pricePerNight) implements Serializable {
+                           Integer pricePerNight) implements Serializable {
     public PostResponse {
     }
 }
